refactor(code): move MCP servers to dedicated feature, fix rail height
- Move McpServersSettings + supporting components/hooks out of
  features/settings into a dedicated features/mcp-servers feature.
- Rename the entrypoint to McpServersView; render it as a real
  top-level view (Flex height="100%") so the installed-rail extends
  the full content height instead of collapsing to its inner content.
- Remove the MCP servers entry from the Settings sidebar (id, title,
  component map, SettingsCategory union) and drop the now-unused
  fullwidth branch in SettingsDialog.

Generated-By: PostHog Code
Task-Id: 3516a98e-d297-4c74-a78e-ed7a96855927
