+++ apps/code/src/renderer/features/mcp-servers/components/parts/ServerDetailView.tsx
@@ -1,4 +1,4 @@
-import { useMcpInstallationTools } from "@features/settings/hooks/useMcpInstallationTools";
+import { useMcpInstallationTools } from "@features/mcp-servers/hooks/useMcpInstallationTools";
 import {
   ArrowClockwise,
   ArrowLeft,
