+++ apps/code/src/renderer/features/mcp-servers/components/parts/ToolPolicyToggle.tsx
