+++ apps/code/src/renderer/features/mcp-servers/components/parts/statusBadge.test.ts
