+++ apps/code/src/renderer/features/mcp-servers/hooks/useMcpInstallationTools.ts
