+++ apps/code/src/renderer/features/settings/components/SettingsDialog.tsx
@@ -22,7 +22,6 @@ import {
   HardDrives,
   Keyboard,
   Palette,
-  Plugs,
   SignOut,
   TrafficSignal,
   TreeStructure,
@@ -37,7 +36,6 @@ import { ClaudeCodeSettings } from "./sections/ClaudeCodeSettings";
 import { CloudEnvironmentsSettings } from "./sections/CloudEnvironmentsSettings";
 import { EnvironmentsSettings } from "./sections/environments/EnvironmentsSettings";
 import { GeneralSettings } from "./sections/GeneralSettings";
-import { McpServersSettings } from "./sections/McpServersSettings";
 import { PersonalizationSettings } from "./sections/PersonalizationSettings";
 import { PlanUsageSettings } from "./sections/PlanUsageSettings";
 import { ShortcutsSettings } from "./sections/ShortcutsSettings";
@@ -51,7 +49,6 @@ interface SidebarItem {
   label: string;
   icon: ReactNode;
   hasChevron?: boolean;
-  fullwidth?: boolean;
 }
 
 const SIDEBAR_ITEMS: SidebarItem[] = [
@@ -75,12 +72,6 @@ const SIDEBAR_ITEMS: SidebarItem[] = [
     icon: <Palette size={16} />,
   },
   { id: "claude-code", label: "Claude Code", icon: <Code size={16} /> },
-  {
-    id: "mcp-servers",
-    label: "MCP servers",
-    icon: <Plugs size={16} />,
-    fullwidth: true,
-  },
   { id: "shortcuts", label: "Shortcuts", icon: <Keyboard size={16} /> },
 
   {
@@ -101,7 +92,6 @@ const CATEGORY_TITLES: Record<SettingsCategory, string> = {
   "cloud-environments": "Cloud environments",
   personalization: "Personalization",
   "claude-code": "Claude Code",
-  "mcp-servers": "MCP Servers",
   shortcuts: "Shortcuts",
 
   signals: "Signals",
@@ -118,7 +108,6 @@ const CATEGORY_COMPONENTS: Record<SettingsCategory, React.ComponentType> = {
   "cloud-environments": CloudEnvironmentsSettings,
   personalization: PersonalizationSettings,
   "claude-code": ClaudeCodeSettings,
-  "mcp-servers": McpServersSettings,
   shortcuts: ShortcutsSettings,
 
   signals: SignalSourcesSettings,
@@ -169,8 +158,6 @@ export function SettingsDialog() {
   }
 
   const ActiveComponent = CATEGORY_COMPONENTS[activeCategory];
-  const activeItem = sidebarItems.find((i) => i.id === activeCategory);
-  const isFullwidth = !!activeItem?.fullwidth;
 
   const initials = user
     ? user.first_name && user.last_name
@@ -278,22 +265,16 @@ export function SettingsDialog() {
               fill="url(#settings-dot-pattern)"
             />
           </svg>
-          {isFullwidth ? (
-            <div className="relative z-[1] flex h-full min-h-0 w-full">
-              <ActiveComponent />
-            </div>
-          ) : (
-            <ScrollArea className="h-full w-full">
-              <Box p="6" mx="auto" className="relative z-[1] max-w-[800px]">
-                <Flex direction="column" gap="4">
-                  <Text className="font-medium text-lg leading-6.5">
-                    {CATEGORY_TITLES[activeCategory]}
-                  </Text>
-                  <ActiveComponent />
-                </Flex>
-              </Box>
-            </ScrollArea>
-          )}
+          <ScrollArea className="h-full w-full">
+            <Box p="6" mx="auto" className="relative z-[1] max-w-[800px]">
+              <Flex direction="column" gap="4">
+                <Text className="font-medium text-lg leading-6.5">
+                  {CATEGORY_TITLES[activeCategory]}
+                </Text>
+                <ActiveComponent />
+              </Flex>
+            </Box>
+          </ScrollArea>
         </div>
       </div>
     </div>
